Use importlib to support Ubuntu 24.04 server
