+++ ansible/plugins/connection/multi_passwd_ssh.py
@@ -1,39 +1,18 @@
 import hashlib
-try:
-    import importlib.util
-    import importlib.machinery
-    use_importlib = True
-except ImportError:
-    import imp
-    use_importlib = False
 import logging
 import os
 
 from functools import wraps
 from ansible.errors import AnsibleAuthenticationFailure, AnsibleConnectionFailure
 from ansible.plugins import connection
-
+from ansible.devutil.testbed import load_source
 
 logger = logging.getLogger(__name__)
 
 
-def load_source(modname, filename):
-    loader = importlib.machinery.SourceFileLoader(modname, filename)
-    spec = importlib.util.spec_from_file_location(modname, filename, loader=loader)
-    module = importlib.util.module_from_spec(spec)
-    # The module is always executed and not cached in sys.modules.
-    # Uncomment the following line to cache the module.
-    # sys.modules[module.__name__] = module
-    loader.exec_module(module)
-    return module
-
-
 # HACK: workaround to import the SSH connection plugin
 _ssh_mod = os.path.join(os.path.dirname(connection.__file__), "ssh.py")
-if use_importlib:
-    _ssh = load_source("_ssh", _ssh_mod)
-else:
-    _ssh = imp.load_source("_ssh", _ssh_mod)
+_ssh = load_source("_ssh", _ssh_mod)
 
 # Use same options as the builtin Ansible SSH plugin
 DOCUMENTATION = _ssh.DOCUMENTATION
